proxy/requirements.txt
fastapi
httpx
uvicorn
grpcio
grpcio-tools
python-jose
bcrypt==4.0.1
passlib
python-jose